sessioh helper, client side request
